requirements.txt
pytest>=8.0.0
playwright>=1.42.0
pytest-playwright>=0.5.0
pandas>=2.2.0
allure-pytest>=2.13.2
pytest-rerunfailures>=13.0
pytest-xdist>=3.5.0